+++ test/test_transformations.py
@@ -86,6 +86,45 @@ def test_copy(some_thr, any_dtype):
     assert diff_is_negligible(output_dev.get(), input_)
 
 
+def test_add_param(some_thr, any_dtype):
+
+    input = get_test_array((1000,), any_dtype)
+    p1 = get_test_array((1,), any_dtype)[0]
+    p2 = get_test_array((1,), any_dtype)[0]
+    input_dev = some_thr.to_device(input)
+    output_dev = some_thr.empty_like(input_dev)
+
+    test = get_test_computation(input_dev)
+    scale = tr.add_param(input_dev, any_dtype)
+
+    test.parameter.input.connect(scale, scale.output, input_prime=scale.input, p1=scale.param)
+    test.parameter.output.connect(scale, scale.input, output_prime=scale.output, p2=scale.param)
+    testc = test.compile(some_thr)
+
+    testc(output_dev, p1, input_dev, p2)
+    assert diff_is_negligible(output_dev.get(), input + p1 + p2)
+
+
+def test_add_const(some_thr, any_dtype):
+
+    input = get_test_array((1000,), any_dtype)
+    p1 = get_test_array((1,), any_dtype)[0]
+    p2 = get_test_array((1,), any_dtype)[0]
+    input_dev = some_thr.to_device(input)
+    output_dev = some_thr.empty_like(input_dev)
+
+    test = get_test_computation(input_dev)
+    scale1 = tr.add_const(input_dev, p1)
+    scale2 = tr.add_const(input_dev, p2)
+
+    test.parameter.input.connect(scale1, scale1.output, input_prime=scale1.input)
+    test.parameter.output.connect(scale2, scale2.input, output_prime=scale2.output)
+    testc = test.compile(some_thr)
+
+    testc(output_dev, input_dev)
+    assert diff_is_negligible(output_dev.get(), input + p1 + p2)
+
+
 def test_scale_param(some_thr, any_dtype):
 
     input = get_test_array((1000,), any_dtype)